This is version 1.0. The results presented in the paper were estimated using the functions as they are in this version of the package except for a few details which would have no impact on the results.
One modification with this commit is an update of the function used to evaluate lagged relationships between risk factors and the probability of new infection. This is to handle cases where there are mutliple test results in a herd on a given month. Duplicates are removed. The first result is taken into account. This situation should be rare.