I wonder if something like this might work to add weeks and years?
This is not fully functional I guess but just an idea if it could
help.

I think the right approach is probably more generalized like the user
submits a vector of break points and you just use those to cut the
dates in the data. But if you are going to use the format approach
maybe this is a step in the right direction